device-management/requirements.txt
# Device Management Microservice Dependencies

# Web Framework
fastapi==0.123.7
uvicorn==0.38.0

# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.11

# Validation
pydantic==2.12.5
email-validator==2.3.0

# HTTP Client
httpx==0.28.1

# Utilities
python-multipart==0.0.20
python-dotenv==1.2.1

# Logging
pyyaml==6.0.3

# Redis Cache
redis==5.2.0

# JWT Authentication
python-jose==3.3.0

# RabbitMQ Messaging
pika==1.3.2

# Testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
